+++ tests/kitchen-sink.test.js
@@ -7,6 +7,9 @@ crosscheck(({ stable, oxide }) => {
       content: [
         {
           raw: html`
+            <div
+              class="line-clamp-2 line-clamp-[33] line-clamp-[var(--line-clamp-variable)] line-clamp-none"
+            ></div>
             <div class="range:text-right multi:text-left"></div>
             <div
               class="container hover:container sm:container md:container text-center sm:text-center md:text-center"
@@ -303,7 +306,7 @@ crosscheck(({ stable, oxide }) => {
         }
         .drop-empty-rules:hover,
         .group:hover .apply-group,
-        .dark .apply-dark-mode {
+        :is(.dark .apply-dark-mode) {
           font-weight: 700;
         }
         .apply-with-existing:hover {
@@ -338,7 +341,7 @@ crosscheck(({ stable, oxide }) => {
         .apply-order-b {
           margin: 1.5rem 1.25rem 1.25rem;
         }
-        .dark .group:hover .apply-dark-group-example-a {
+        :is(.dark .group:hover .apply-dark-group-example-a) {
           --tw-bg-opacity: 1;
           background-color: rgb(34 197 94 / var(--tw-bg-opacity));
         }
@@ -431,6 +434,27 @@ crosscheck(({ stable, oxide }) => {
         }
         .mt-6 {
           margin-top: 1.5rem;
+        }
+        .line-clamp-2 {
+          -webkit-line-clamp: 2;
+          -webkit-box-orient: vertical;
+          display: -webkit-box;
+          overflow: hidden;
+        }
+        .line-clamp-\[33\] {
+          -webkit-line-clamp: 33;
+          -webkit-box-orient: vertical;
+          display: -webkit-box;
+          overflow: hidden;
+        }
+        .line-clamp-\[var\(--line-clamp-variable\)\] {
+          -webkit-line-clamp: var(--line-clamp-variable);
+          -webkit-box-orient: vertical;
+          display: -webkit-box;
+          overflow: hidden;
+        }
+        .line-clamp-none {
+          -webkit-line-clamp: unset;
         }
         .scale-50 {
           --tw-scale-x: 0.5;
@@ -715,7 +739,7 @@ crosscheck(({ stable, oxide }) => {
             transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
           }
         }
-        .dark .dark\:custom-util {
+        :is(.dark .dark\:custom-util) {
           background: #abcdef;
         }
         @media (min-width: 640px) {
@@ -762,7 +786,7 @@ crosscheck(({ stable, oxide }) => {
               transition-duration: 0.15s;
               transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
             }
-            .dark .md\:dark\:motion-safe\:foo\:active\:custom-util:active {
+            :is(.dark .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
               background: #abcdef !important;
             }
           }
@@ -849,7 +873,7 @@ crosscheck(({ stable, oxide }) => {
         }
         .drop-empty-rules:hover,
         .group:hover .apply-group,
-        .dark .apply-dark-mode {
+        :is(.dark .apply-dark-mode) {
           font-weight: 700;
         }
         .apply-with-existing:hover {
@@ -883,7 +907,7 @@ crosscheck(({ stable, oxide }) => {
         .apply-order-b {
           margin: 1.5rem 1.25rem 1.25rem;
         }
-        .dark .group:hover .apply-dark-group-example-a {
+        :is(.dark .group:hover .apply-dark-group-example-a) {
           background-color: #22c55e;
         }
         @media (min-width: 640px) {
@@ -975,6 +999,27 @@ crosscheck(({ stable, oxide }) => {
         }
         .mt-6 {
           margin-top: 1.5rem;
+        }
+        .line-clamp-2 {
+          -webkit-line-clamp: 2;
+          -webkit-box-orient: vertical;
+          display: -webkit-box;
+          overflow: hidden;
+        }
+        .line-clamp-\[33\] {
+          -webkit-line-clamp: 33;
+          -webkit-box-orient: vertical;
+          display: -webkit-box;
+          overflow: hidden;
+        }
+        .line-clamp-\[var\(--line-clamp-variable\)\] {
+          -webkit-line-clamp: var(--line-clamp-variable);
+          -webkit-box-orient: vertical;
+          display: -webkit-box;
+          overflow: hidden;
+        }
+        .line-clamp-none {
+          -webkit-line-clamp: unset;
         }
         .scale-50 {
           --tw-scale-x: 0.5;
@@ -1250,7 +1295,7 @@ crosscheck(({ stable, oxide }) => {
             transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
           }
         }
-        .dark .dark\:custom-util {
+        :is(.dark .dark\:custom-util) {
           background: #abcdef;
         }
         @media (min-width: 640px) {
@@ -1297,7 +1342,7 @@ crosscheck(({ stable, oxide }) => {
               transition-duration: 0.15s;
               transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
             }
-            .dark .md\:dark\:motion-safe\:foo\:active\:custom-util:active {
+            :is(.dark .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
               background: #abcdef !important;
             }
           }