+++ tests/variants.test.js
@@ -1036,7 +1036,7 @@ crosscheck(({ stable, oxide }) => {
     })
   })
 
-  test('variants with slashes suspport modifiers', () => {
+  test('variants with slashes support modifiers', () => {
     let config = {
       content: [
         {
@@ -1125,6 +1125,30 @@ crosscheck(({ stable, oxide }) => {
       }
       .\[\&\:\:-webkit-scrollbar\:hover\]\:underline::-webkit-scrollbar:hover {
         text-decoration-line: underline;
+      }
+    `)
+  })
+
+  test('stacking dark and rtl variants', async () => {
+    let config = {
+      darkMode: 'class',
+      content: [
+        {
+          raw: html` <div class="dark:rtl:italic" /> `,
+        },
+      ],
+      corePlugins: { preflight: false },
+    }
+
+    let input = css`
+      @tailwind utilities;
+    `
+
+    let result = await run(input, config)
+
+    expect(result.css).toMatchFormattedCss(css`
+      :is(.dark :is([dir='rtl'] .dark\:rtl\:italic)) {
+        font-style: italic;
       }
     `)
   })